moved graphql utils, added function to convert job states
